requirements.txt
fastapi
uvicorn
requests
cloudscraper
psycopg2-binary
pymongo
recipe-scrapers
python-dotenv
